Refactor to actually use a factory for subjects
