+++ src/test/java/homeschooltracker/users/StudentTest.java
@@ -1,7 +1,6 @@
 package homeschooltracker.users;
 
-import homeschooltracker.content.Subject;
-import homeschooltracker.content.SubjectGrade1Reading2;
+import homeschooltracker.content.subjects.SubjectGrade1Reading2;
 import junit.framework.TestCase;
 
 public class StudentTest extends TestCase {
@@ -23,10 +22,10 @@ public class StudentTest extends TestCase {
 
     public void testGetSubjectAtPosition() {
         SubjectGrade1Reading2 subject1 = new SubjectGrade1Reading2();
-        Subject subject2 = new Subject();
         student.addSubject(subject1);
-        student.addSubject(subject2);
+        // Add second subject
+        student.addSubject(subject1);
 
-        assertEquals("Unnamed Subject", student.getSubjectAtPosition(1).getName());
+        assertEquals("Reading 2, Grade 1", student.getSubjectAtPosition(1).getName());
     }
 }
